Cygwin: FIFO: designate one reader as owner
Among all the open readers of a FIFO, one is declared to be the owner.
This is the only reader that listens for client connections, and it is
the only one that can read.

Remove the 'first' argument from create_pipe_instance, which is not
needed.  Moreover, with the coming changes to support multiple
readers, the previous criterion for setting 'first' (nhandlers == 0)
may not remain valid.