Cygwin: FIFO: improve the retry loop when opening a writer
When a writer tries to connect to a pipe instance, NtOpenFile can
occasionally fail with STATUS_OBJECT_NAME_NOT_FOUND.  See

  https://sourceware.org/pipermail/cygwin/2020-April/244373.html

Apparently there can be a delay between the time the Windows named
pipe is created and the time it shows up in the file system.

Build this possibility into the retry loop in open.

Also avoid a WFSO call in the nonblocking case.